mise à jour de la suppression des dossiers pour ne pas supprimer le dossier racine

+++ SoundLink_Server/soundlink_service/src/main/java/soundlink/service/websocket/handler/impl/MusicIntegrationExecutor.java
@@ -136,7 +136,9 @@ public class MusicIntegrationExecutor extends WebSocketExecutor implements IMusi
             }
         }
         // delete input directory
-        directory.delete();
+        if (!directory.equals(musicsInputFolder)) {
+            directory.delete();
+        }
     }
 
     private void processFile(File musicFile) {
